optimize UpdatePredictionCache
